booking-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
playwright==1.44.0
python-telegram-bot==21.3
pydantic==2.7.0
python-dotenv==1.0.1
pillow==10.3.0
requests==2.32.3
groq==0.9.0
